# gen/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "contact": {
            "email": "support@codespace.dev",
            "name": "Codespace Operator Support",
            "url": "https://github.com/codespace-operator/codespace-operator"
        },
        "description": "REST API for managing Codespace sessions with RBAC and multi-provider authentication",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "termsOfService": "https://github.com/codespace-operator/codespace-operator/blob/main/TERMS.md",
        "title": "Codespace Operator API",
        "version": "2.0.0"
    },
    "paths": {
        "/api/v1/admin/rbac/reload": {
            "post": {
                "description": "Force reload of RBAC policies (requires admin privileges)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "Reload RBAC (Admin)",
                "tags": [
                    "admin"
                ]
            }
        },
...
        "/api/v1/admin/system/info": {
            "get": {
                "description": "Check system information (requires admin privileges)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "System Info",
                "tags": [
                    "admin"
                ]
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "description": "Get list of users in the system (requires admin privileges)",
                "responses": {
                    "200": {
                        "description": "User list with admin info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "List users (Admin)",
                "tags": [
                    "admin"
                ]
            }
        },
...
    },
...
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer token authentication (JWT)",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            },
            "CookieAuth": {
                "description": "Session cookie authentication",
                "in": "cookie",
                "name": "codespace_session",
                "type": "apiKey"
            }
        },
        "schemas": {
...
            "internal_server.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "Invalid request",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}